Fix
- 유효성 검사 문구를 수정하였습니다(EditNickName 스크린)
- MyPage 의 유효성 검사에 Alert 기능을 추가하였습니다. ( 확인버튼 누를시 로그아웃/탈퇴 기능 )
- MyPage의 디데이 계산 로직을 커스텀훅으로 리팩토링하였습니다.

by ddoqi
